+++ layers/CMSSchema/packages/post-mutations/src/TypeResolvers/ObjectType/PostUpdateMutationPayloadObjectTypeResolver.php
@@ -13,6 +13,6 @@ class PostUpdateMutationPayloadObjectTypeResolver extends AbstractPostMutationPa
 
     public function getTypeDescription(): ?string
     {
-        return $this->__('Payload of executing an update mutation on a post', 'post-mutations');
+        return $this->__('Payload of executing an update nested mutation on a post', 'post-mutations');
     }
 }