+++ layers/CMSSchema/packages/post-mutations/src/TypeResolvers/ObjectType/RootUpdatePostMutationPayloadObjectTypeResolver.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPCMSSchema\PostMutations\TypeResolvers\ObjectType;
+
+class RootUpdatePostMutationPayloadObjectTypeResolver extends AbstractPostMutationPayloadObjectTypeResolver
+{
+    public function getTypeName(): string
+    {
+        return 'RootUpdatePostMutationPayload';
+    }
+
+    public function getTypeDescription(): ?string
+    {
+        return $this->__('Payload of executing an update mutation on a post', 'post-mutations');
+    }
+}